@@ -1,20 +1,25 @@
 {
   "name": "the-gargantuan-frontend",
   "version": "1.6.0",
-  "description": "Frontend application for The Gargantuan project.",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "marked": "^5.0.1",
-    "axios": "^1.6.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
+    "marked": "^12.0.1"
   },
   "devDependencies": {
-    "vite": "^4.4.9",
-    "@vitejs/plugin-react": "^4.0.0"
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.3.5",
+    "vite": "^4.4.9"
+  },
+  "engines": {
+    "node": ">=18"
   }
 }
