@@ -40,13 +40,10 @@ export default function ManagePosts({ token, toast }) {
     load();
   }, []);
 
-  // Toggle selection of all rows.  When the master checkbox is checked all
-  // published posts and drafts will be marked as selected; when unchecked
-  // they will all be cleared.
-  function toggleAll(e) {
+  function toggleAll(e, which) {
     const checked = e.target.checked;
     const next = { ...selected };
-    const list = [...(posts || []), ...(drafts || [])];
+    const list = which === 'draft' ? drafts : posts;
     if (checked) list.forEach((p) => { next[p.id] = true; });
     else list.forEach((p) => { next[p.id] = false; });
     setSelected(next);
@@ -129,12 +126,18 @@ export default function ManagePosts({ token, toast }) {
 
   // Compose rows for both published posts and drafts.
   const rows = useMemo(() => {
-    // Combine published posts and drafts into a single array.  Include a
-    // snippet of the body for quick preview in the table.  The snippet is
-    // truncated to 60 characters for compact display.
     const all = [...(posts || []), ...(drafts || [])];
     return all.map((p, i) => {
-      const snippet = p.body ? (p.body.length > 60 ? `${p.body.slice(0, 60)}…` : p.body) : '';
+      // Build a small text snippet for preview if no image is provided.
+      let snippet = '';
+      if (p.body) {
+        // Strip markdown syntax and truncate.
+        const plain = p.body.replace(/\!\[[^\]]*\]\([^)]*\)/g, '') // remove images
+          .replace(/\[[^\]]*\]\([^)]*\)/g, '') // remove links
+          .replace(/[`_*#>-]/g, '') // remove md syntax
+          .trim();
+        snippet = plain.substring(0, 60) + (plain.length > 60 ? '…' : '');
+      }
       return {
         key: i,
         id: p.id,
@@ -143,31 +146,11 @@ export default function ManagePosts({ token, toast }) {
         type: p.type,
         draft: p.draft,
         url: p.playUrl || p.audioUrl || p.videoUrl || '',
+        imageUrl: p.imageUrl || '',
         snippet,
       };
     });
   }, [posts, drafts]);
-
-  // Toggle the draft status of a post.  If the post is currently a draft
-  // (`true`) then toggling will publish it (`false`), and vice versa.  After
-  // saving the draft flag the posts list is refreshed.
-  async function toggleDraft(id, currentDraft) {
-    const res = await fetch(`${API_BASE}/api/posts/${encodeURIComponent(id)}`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json', 'x-admin-token': token },
-      body: JSON.stringify({ draft: !currentDraft }),
-    });
-    if (res.status === 401) {
-      toast?.show('Unauthorized (toggle draft).', 'error');
-      return;
-    }
-    if (!res.ok) {
-      toast?.show('Failed to toggle draft', 'error');
-      return;
-    }
-    toast?.show(currentDraft ? 'Published' : 'Unpublished', 'ok');
-    load();
-  }
 
   return (
     <div className="bg-white border border-[#dcdcdc] rounded-lg p-4 sm:p-6">
@@ -187,12 +170,12 @@ export default function ManagePosts({ token, toast }) {
           <table className="w-full text-sm">
             <thead className="text-left text-[#555]">
               <tr>
-                <th className="py-2 pr-4"><input type="checkbox" onChange={toggleAll} /></th>
+                <th className="py-2 pr-4"><input type="checkbox" onChange={(e) => toggleAll(e, 'all')} /></th>
                 <th className="py-2 pr-4">Title</th>
-                <th className="py-2 pr-4">Preview</th>
                 <th className="py-2 pr-4">Date</th>
                 <th className="py-2 pr-4">Type</th>
                 <th className="py-2 pr-4">Draft</th>
+                <th className="py-2 pr-4">Preview</th>
                 <th className="py-2 pr-4">Toggle</th>
                 <th className="py-2">Actions</th>
               </tr>
@@ -206,7 +189,9 @@ export default function ManagePosts({ token, toast }) {
                   onToggle={toggleOne}
                   onDelete={delOne}
                   onEdit={startEdit}
-                  onToggleDraft={toggleDraft}
+                  toast={toast}
+                  token={token}
+                  reload={load}
                 />
               ))}
             </tbody>
@@ -243,26 +228,48 @@ export default function ManagePosts({ token, toast }) {
   );
 }
 
-function Row({ row, checked, onToggle, onDelete, onEdit, onToggleDraft }) {
-  // Apply a yellow background to draft rows to help users differentiate drafts
-  // from published posts.  Each row displays a preview snippet, the publish
-  // state and a button to toggle draft status.
+function Row({ row, checked, onToggle, onDelete, onEdit, toast, token, reload }) {
+  // Toggle draft status.  Calls PATCH /api/posts/:id with the inverted draft flag.
+  async function handleToggle() {
+    try {
+      const res = await fetch(`${API_BASE}/api/posts/${encodeURIComponent(row.id)}`, {
+        method: 'PATCH',
+        headers: { 'Content-Type': 'application/json', 'x-admin-token': token },
+        body: JSON.stringify({ draft: !row.draft }),
+      });
+      if (res.status === 401) {
+        toast?.show('Unauthorized (toggle).', 'error');
+        return;
+      }
+      if (!res.ok) {
+        toast?.show('Toggle failed', 'error');
+        return;
+      }
+      toast?.show(row.draft ? 'Published' : 'Unpublished', 'ok');
+      reload?.();
+    } catch (err) {
+      toast?.show(err.message || 'Toggle failed', 'error');
+    }
+  }
   return (
-    <tr className={`border-t border-[#eee] align-top ${row.draft ? 'bg-yellow-50' : ''}`}>
+    <tr className="border-t border-[#eee]">
       <td className="py-2 pr-4"><input type="checkbox" checked={checked} onChange={(e) => onToggle(row.id, e.target.checked)} /></td>
-      <td className="py-2 pr-4 whitespace-nowrap">{row.title}</td>
-      <td className="py-2 pr-4 max-w-xs truncate" title={row.snippet}>{row.snippet}</td>
-      <td className="py-2 pr-4 whitespace-nowrap">{row.date ? new Date(row.date).toLocaleString() : ''}</td>
-      <td className="py-2 pr-4 uppercase whitespace-nowrap">{row.type}</td>
-      <td className="py-2 pr-4 whitespace-nowrap">{row.draft ? 'Yes' : 'No'}</td>
-      <td className="py-2 pr-4 whitespace-nowrap">
-        <button
-          onClick={() => onToggleDraft(row.id, row.draft)}
-          className={`px-3 py-1 rounded text-white text-xs ${row.draft ? 'bg-green-600 hover:bg-green-700' : 'bg-yellow-600 hover:bg-yellow-700'}`}
-        >{row.draft ? 'Publish' : 'Unpublish'}</button>
+      <td className="py-2 pr-4 max-w-xs truncate" title={row.title}>{row.title}</td>
+      <td className="py-2 pr-4">{row.date ? new Date(row.date).toLocaleString() : ''}</td>
+      <td className="py-2 pr-4 uppercase">{row.type}</td>
+      <td className="py-2 pr-4">{row.draft ? 'Yes' : 'No'}</td>
+      <td className="py-2 pr-4">
+        {row.imageUrl ? (
+          <img src={row.imageUrl} alt={row.title} className="w-14 h-10 object-cover rounded" />
+        ) : (
+          <span className="text-xs text-[#666]">{row.snippet || ''}</span>
+        )}
+      </td>
+      <td className="py-2 pr-4">
+        <button onClick={handleToggle} className={`px-3 py-1 rounded text-white text-xs ${row.draft ? 'bg-[#052962]' : 'bg-[#c70000]'}`}>{row.draft ? 'Publish' : 'Unpublish'}</button>
       </td>
       <td className="py-2">
-        <div className="flex gap-2 flex-wrap">
+        <div className="flex flex-wrap gap-2">
           <button onClick={() => onEdit(row)} className="px-3 py-1 bg-[#052962] text-white rounded text-xs">Edit</button>
           {row.url && <a href={row.url} target="_blank" className="px-3 py-1 border border-[#dcdcdc] rounded text-xs">Open</a>}
           <button onClick={() => onDelete(row.id)} className="px-3 py-1 bg-[#c70000] text-white rounded text-xs">Delete</button>
