@@ -1,16 +1,5 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-// Export Vite configuration.  When deployed on Vercel the environment
-// variable VITE_API_BASE must be set to the backend URL.  During
-// development this can be overridden via .env.local.
-export default defineConfig({
-  plugins: [react()],
-  build: {
-    outDir: 'dist',
-    sourcemap: false,
-  },
-  server: {
-    port: 5173,
-  },
-});
+import { defineConfig } from 'vite'
+// We omit the React plugin to avoid pulling in private dependencies which
+// may be blocked on certain registries.  Vite's built-in esbuild loader
+// will still transform JSX using the automatic runtime.
+export default defineConfig({})
